Fix queued requests not taking updated credentials
Queued requests are re-sent in a way that "$http.default.headers" are ignored, which causes updated credentials never been picked up.

The fix extends the header of a queued request with the latest "$http.default.headers" before retrying so that custom headers in the original request (if applicable) will be retained and the latest auth headers attatched.

This fixes #5.